[Refactor] index.ts modify, apply
