+++ src/components/home/index.ts
@@ -1,0 +1,1 @@
+export { default as LedgerList } from './LedgerList';
