+++ src/components/ledgers/LedgerItem.tsx
@@ -2,14 +2,14 @@ import React from 'react';
 import styled from 'styled-components';
 import { theme } from '@/styles';
 import { amountTostring } from '@/lib/utils/string';
-import { P } from '../common';
-import { ArchiveType, TagType } from '@/types/tag';
+import { HashTagGroup, P } from '../common';
+import { ArchiveType, TagType } from '@/types';
 
 interface ILedgerItemProps {
   title: string;
   date: string;
   amount: number;
-  tags?: TagType[];
+  tags: TagType[];
   category: ArchiveType;
 }
 
@@ -24,11 +24,16 @@ export default function LedgerItem(props: ILedgerItemProps) {
   };
   return (
     <Wrapper categoryID={category.archiveTypeID.toString()}>
-      <DTWrapper>
-        <Date>{toShortDate(date)}</Date>
-        <Title>{title}</Title>
-      </DTWrapper>
-      <Amount>{amountTostring(amount)}</Amount>
+      <BasicWrapper>
+        <DTWrapper>
+          <Date>{toShortDate(date)}</Date>
+          <Title>{title}</Title>
+        </DTWrapper>
+        <Amount>{amountTostring(amount)}</Amount>
+      </BasicWrapper>
+      <ExtendWrapper isOpen>
+        <HashTagGroup tagList={tags ?? []} handleSelect={() => {}} />
+      </ExtendWrapper>
     </Wrapper>
   );
 }
@@ -37,13 +42,18 @@ const Wrapper = styled.div<{ categoryID: string }>`
   padding: 0.9rem 1.5rem;
   background-color: ${(props) => theme.color.LEDGER_BACKGROUND[props.categoryID]};
   width: 100%;
-  height: 4.2rem;
+  /* height: 4.2rem; */
+  display: flex;
+  flex-direction: column;
+  border-radius: 10px;
+  margin-bottom: 0.5rem;
+`;
+
+const BasicWrapper = styled.div`
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
-  border-radius: 10px;
-  margin-bottom: 0.5rem;
 `;
 
 const DTWrapper = styled.div`
@@ -66,3 +76,7 @@ const Title = styled(P).attrs({
 const Amount = styled(P).attrs({
   fontWeight: theme.font.fontWeight.regular,
 })``;
+
+const ExtendWrapper = styled.div<{ isOpen: boolean }>`
+  max-height: ${(props) => (props.isOpen ? 'none' : 0)};
+`;
