+++ src/styles/index.ts
@@ -1,4 +1,2 @@
-import { GlobalStyle } from "./global";
-import { theme } from "./theme";
-
-export {GlobalStyle,theme};
+export * from './global';
+export * from './theme';
